+++ packages/cms/lib/modules/accordeon-widgets/index.js
@@ -47,10 +47,13 @@ module.exports = {
     self.load = (req, widgets, callback) => {
       widgets.forEach((widget) => {
         if (widget.containerStyles) {
-          const containerId = widget._id;
+          const containerId = self.apos.utils.generateId();
           widget.containerId = containerId;
           widget.formattedContainerStyles = styleSchema.format(containerId, widget.containerStyles);
         }
+
+        widget.cssHelperClassesString = widget.cssHelperClasses ? widget.cssHelperClasses.join(' ') : '';
+
       });
 
       return superLoad(req, widgets, callback);