+++ packages/cms/lib/modules/counter-widgets/index.js
@@ -96,10 +96,12 @@ module.exports = {
           }
 
           if (widget.containerStyles) {
-            const containerId = widget._id;
+            const containerId = self.apos.utils.generateId();
             widget.containerId = containerId;
             widget.formattedContainerStyles = styleSchema.format(containerId, widget.containerStyles);
           }
+          widget.cssHelperClassesString = widget.cssHelperClasses ? widget.cssHelperClasses.join(' ') : '';
+
           widget.siteId = req.data.global.siteId;
         });