+++ packages/cms/lib/modules/ideas-on-map-widgets/index.js
@@ -25,7 +25,7 @@ module.exports = {
       {
         name: 'map',
         label: 'Kaart',
-        fields: ['mapVariant', 'mapAutoZoomAndCenter', 'mapLocationIcon', 'mapClustering', 'mapMaxClusterRadius', 'canSelectLocation' ]
+        fields: ['mapVariant', 'mapTilesUrl', 'mapTilesSubdomains', 'mapAutoZoomAndCenter', 'mapLocationIcon', 'mapClustering', 'mapMaxClusterRadius', 'canSelectLocation', 'onMarkerClickAction' ]
       },
       {
         name: 'content',
@@ -72,9 +72,10 @@ module.exports = {
 			  widget.config = JSON.stringify(createConfig(widget, req.data, req.session.jwt, self.apos.settings.getOption(req, 'apiUrl'), req.data.siteUrl + '/oauth/login?returnTo=' + encodeURIComponent(req.url) ));
         widget.openstadComponentsUrl = openstadComponentsUrl;
 
-        const containerId = widget._id;
+        const containerId = self.apos.utils.generateId();
         widget.containerId = containerId;
-        widget.formattedContainerStyles = styleSchema.format(containerId, widget.containerStyles);
+              widget.cssHelperClassesString = widget.cssHelperClasses ? widget.cssHelperClasses.join(' ') : '';
+              widget.formattedContainerStyles = styleSchema.format(containerId, widget.containerStyles);
 
       });