+++ packages/cms/lib/modules/previous-next-button-block-widgets/index.js
@@ -49,9 +49,10 @@ module.exports = {
           nextLabel: widget.nextLabel,
         });
         widget.openstadComponentsUrl = openstadComponentsUrl;
-        const containerId = widget._id;
+        const containerId = self.apos.utils.generateId();
         widget.containerId = containerId;
-        widget.formattedContainerStyles = styleSchema.format(containerId, widget.containerStyles);
+                widget.cssHelperClassesString = widget.cssHelperClasses ? widget.cssHelperClasses.join(' ') : '';
+                widget.formattedContainerStyles = styleSchema.format(containerId, widget.containerStyles);
 			});
 
 			return superLoad(req, widgets, next);