+++ packages/cms/lib/modules/resource-overview-widgets/index.js
@@ -110,7 +110,7 @@ module.exports = {
                 widget.themes = req.data.global.themes;
                 widget.areas = req.data.global.areas;
 
-                const containerId = widget._id;
+                const containerId = self.apos.utils.generateId();
                 widget.containerId = containerId;
 
                 widget.parseDateToTime = (date) => {
@@ -160,6 +160,9 @@ module.exports = {
 
                     widget.themes = includeThemes.length > 0 ? widget.themes.filter(theme => includeThemes.indexOf(theme.value) > -1) : widget.themes;
                 }
+
+                widget.cssHelperClassesString = widget.cssHelperClasses ? widget.cssHelperClasses.join(' ') : '';
+
 
                 if (widget.containerStyles) {
                     widget.formattedContainerStyles = styleSchema.format(containerId, widget.containerStyles);