+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Remove warning not the right budgeting type
 * Better export of choices-guide results
 * Move google static map image to interactive map in Participatory Budgeting Widget
+* Added configuration of idea types to global settings
 
 ## v0.14.0 (2020-03-17)
 * Load files with static middleware instead of sendFile, fallback to ApostropheCMS in case a file is not found